#394 Changed hitbox state machine. --snapshot

+++ petblocks-bukkit-plugin/petblocks-bukkit-nms-111R1/src/main/java/com/github/shynixn/petblocks/bukkit/logic/business/nms/v1_11_R1/NMSPetArmorstand.kt
@@ -77,7 +77,8 @@ class NMSPetArmorstand(owner: Player, val petMeta: PetMeta) : EntityArmorStand((
 
         val location = owner.location
         val mcWorld = (location.world as CraftWorld).handle
-        val position = PositionEntity(location.x, location.y, location.z, location.yaw.toDouble(), location.pitch.toDouble(), location.world.name).relativeFront(3.0)
+        val position =
+            PositionEntity(location.x, location.y, location.z, location.yaw.toDouble(), location.pitch.toDouble(), location.world.name).relativeFront(3.0)
 
         this.setPositionRotation(position.x, position.y, position.z, location.yaw, location.pitch)
         mcWorld.addEntity(this, CreatureSpawnEvent.SpawnReason.CUSTOM)
@@ -93,7 +94,9 @@ class NMSPetArmorstand(owner: Player, val petMeta: PetMeta) : EntityArmorStand((
      * Spawns a new hitbox
      */
     private fun spawnHitBox() {
-        if (internalHitBox != null) {
+        val shouldDeleteHitBox = shouldDeleteHitBox()
+
+        if (shouldDeleteHitBox && internalHitBox != null) {
             (internalHitBox!!.bukkitEntity as EntityPetProxy).deleteFromWorld()
             internalHitBox = null
             proxy.changeHitBox(internalHitBox)
@@ -139,8 +142,11 @@ class NMSPetArmorstand(owner: Player, val petMeta: PetMeta) : EntityArmorStand((
         val flyingAi = petMeta.aiGoals.firstOrNull { a -> a is AIFlying }
 
         if (flyingAi != null) {
-            internalHitBox = NMSPetBat(this, getBukkitEntity().location)
-            proxy.changeHitBox(internalHitBox!!.bukkitEntity as LivingEntity)
+            if (internalHitBox == null) {
+                internalHitBox = NMSPetBat(this, getBukkitEntity().location)
+                proxy.changeHitBox(internalHitBox!!.bukkitEntity as LivingEntity)
+            }
+
             val aiGoals = aiService.convertPetAiBasesToPathfinders(proxy, petMeta.aiGoals)
             (internalHitBox as NMSPetBat).applyPathfinders(aiGoals)
             applyNBTTagToHitBox(internalHitBox!!)
@@ -150,16 +156,22 @@ class NMSPetArmorstand(owner: Player, val petMeta: PetMeta) : EntityArmorStand((
         val hoppingAi = petMeta.aiGoals.firstOrNull { a -> a is AIHopping }
 
         if (hoppingAi != null) {
-            internalHitBox = NMSPetRabbit(this, getBukkitEntity().location)
-            proxy.changeHitBox(internalHitBox!!.bukkitEntity as LivingEntity)
+            if (internalHitBox == null) {
+                internalHitBox = NMSPetRabbit(this, getBukkitEntity().location)
+                proxy.changeHitBox(internalHitBox!!.bukkitEntity as LivingEntity)
+            }
+
             val aiGoals = aiService.convertPetAiBasesToPathfinders(proxy, petMeta.aiGoals)
             (internalHitBox as NMSPetRabbit).applyPathfinders(aiGoals)
             applyNBTTagToHitBox(internalHitBox!!)
             return
         }
 
-        internalHitBox = NMSPetVillager(this, getBukkitEntity().location)
-        proxy.changeHitBox(internalHitBox!!.bukkitEntity as LivingEntity)
+        if (internalHitBox == null) {
+            internalHitBox = NMSPetVillager(this, getBukkitEntity().location)
+            proxy.changeHitBox(internalHitBox!!.bukkitEntity as LivingEntity)
+        }
+
         val aiGoals = aiService.convertPetAiBasesToPathfinders(proxy, petMeta.aiGoals)
         (internalHitBox as NMSPetVillager).applyPathfinders(aiGoals)
         applyNBTTagToHitBox(internalHitBox!!)
@@ -445,6 +457,29 @@ class NMSPetArmorstand(owner: Player, val petMeta: PetMeta) : EntityArmorStand((
                 rootCompoundMap[key] = parsedCompoundMap[key]
             }
         }
+    }
+
+    /**
+     * Should the hitbox of the armorstand be deleted.
+     */
+    private fun shouldDeleteHitBox(): Boolean {
+        val hasEmptyHitBoxAi = petMeta.aiGoals.firstOrNull { a -> a is AIGroundRiding || a is AIFlyRiding || a is AIWearing } != null
+
+        if (hasEmptyHitBoxAi) {
+            return true
+        }
+
+        if (internalHitBox != null) {
+            if (internalHitBox is NMSPetVillager && petMeta.aiGoals.firstOrNull { a -> a is AIWalking } == null) {
+                return true
+            } else if (internalHitBox is NMSPetRabbit && petMeta.aiGoals.firstOrNull { a -> a is AIHopping } == null) {
+                return true
+            } else if (internalHitBox is NMSPetBat && petMeta.aiGoals.firstOrNull { a -> a is AIFlying } == null) {
+                return true
+            }
+        }
+
+        return false
     }
 
     /**
